feat: 공통 타입 type파일로 분리
